Microservices/Complex_Microservices/place_bid/requirements.txt
flask_sqlalchemy
flask_cors
pika
pyodbc
flask_socketio
requests==2.31.0